+++ src/components/BatchAllocations/hooks/index.js
@@ -1,5 +1,6 @@
 export { useBatchAllocation } from './useBatchAllocation';
 export { useBatchAllocationLogs } from './useBatchAllocationLogs';
 export { useBatchAllocationFormatter } from './useBatchAllocationFormatter';
+export { useBatchAllocationFormHandler } from './useBatchAllocationFormHandler';
 export { useBatchAllocationMutation } from './useBatchAllocationMutation';
 export { useSourceData } from './useSourceData';